Move elements of `common` into `utils` modules
