+++ agents/world/CountryAgentRunner.java
@@ -3,8 +3,8 @@ package grakn.simulation.agents.world;
 import grakn.simulation.agents.base.Agent;
 import grakn.simulation.agents.base.IterationContext;
 import grakn.simulation.agents.base.AgentRunner;
-import grakn.simulation.common.RandomSource;
-import grakn.simulation.common.Tracker;
+import grakn.simulation.utils.RandomSource;
+import grakn.simulation.agents.utils.Tracker;
 import grakn.simulation.world.World;
 
 import java.util.List;
