+++ utils/BUILD
@@ -1,5 +1,5 @@
 java_library(
-    name = "common",
+    name = "utils",
     srcs = glob(["*.java"]),
     deps = [
         "@maven//:org_slf4j_slf4j_api",
